Change form ignore attributes to retain synonyms
In older versions of formalchemy, synonyms worked. This commit returns
a small amount of functionality that seems to have been dropped at some
point. The original 'ignore_keys' stanza had a special call out for
synonym properties to include. This has the effect of hiding the
properties that are the original target of the synonym but exposing
the synonym property.